requirements.txt
playwrigh == 1.29.1
pytest == 7.2.0